+++ app/builders/dnd2024_character/backgrounds/scribe_builder.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Dnd2024Character
+  module Backgrounds
+    class ScribeBuilder
+      def call(result:)
+        result[:selected_skills] = { investigation: 1, perception: 1 }
+
+        result
+      end
+    end
+  end
+end