+++ app/builders/dnd2024_character/backgrounds/soldier_builder.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Dnd2024Character
+  module Backgrounds
+    class SoldierBuilder
+      def call(result:)
+        result[:selected_feats] = ['savage_attacker']
+        result[:selected_skills] = { athletics: 1, intimidation: 1 }
+
+        result
+      end
+    end
+  end
+end